Rename "start_ingest" to "request_ingest"
