+++ archive/archive_request_ingest/src/test_archive_request_ingest.py
@@ -5,7 +5,7 @@ from uuid import UUID
 
 import pytest
 
-import archive_start_ingest as start_ingest
+import archive_request_ingest as request_ingest
 
 
 TABLE_NAME = 'archive-storage-progress-table'
@@ -14,7 +14,7 @@ TABLE_NAME = 'archive-storage-progress-table'
 def test_post_sends_location_to_sns(sns_client, topic_arn):
     request = ingest_request(upload_url='s3://wellcomecollection-assets-archive-ingest/test-bag.zip')
 
-    response = start_ingest.main(event=request, sns_client=sns_client)
+    response = request_ingest.main(event=request, sns_client=sns_client)
 
     id = str(UUID(response['id']))
     assert id
@@ -36,7 +36,7 @@ def test_sends_request_to_sns_with_callback(sns_client, topic_arn):
     request = ingest_request(upload_url='s3://wellcomecollection-assets-archive-ingest/test-bag.zip',
                              callback_url='https://workflow.wellcomecollection.org/callback?id=b1234567')
 
-    response = start_ingest.main(event=request, sns_client=sns_client)
+    response = request_ingest.main(event=request, sns_client=sns_client)
 
     actual_id = str(UUID(response['id']))
     assert actual_id
@@ -57,7 +57,7 @@ def test_invalid_url_fails(sns_client, topic_arn):
     request = ingest_request('invalidUrl')
 
     with pytest.raises(ValueError, match="\[BadRequest\] Unrecognised url scheme: invalid"):
-        start_ingest.main(event=request, sns_client=sns_client)
+        request_ingest.main(event=request, sns_client=sns_client)
 
     assert len(sns_client.list_messages()) == 0
 
@@ -70,7 +70,7 @@ def test_missing_url_fails(sns_client, topic_arn):
 
     with pytest.raises(KeyError,
                        match="\[BadRequest\] Invalid request missing 'uploadUrl' in {'unknownKey': 'aValue'}"):
-        start_ingest.main(event=request, sns_client=sns_client)
+        request_ingest.main(event=request, sns_client=sns_client)
 
     assert len(sns_client.list_messages()) == 0
 
@@ -82,7 +82,7 @@ def test_invalid_json_fails(sns_client, topic_arn):
     }
 
     with pytest.raises(TypeError, match="\[BadRequest\] Invalid request not json: not_json"):
-        start_ingest.main(event=request, sns_client=sns_client)
+        request_ingest.main(event=request, sns_client=sns_client)
 
     assert len(sns_client.list_messages()) == 0
 
@@ -93,7 +93,7 @@ def test_throws_valueerror_if_called_with_get_event():
     }
 
     with pytest.raises(ValueError, match='Expected request_method=POST'):
-        start_ingest.main(event=event)
+        request_ingest.main(event=event)
 
 
 def test_throws_keyerror_if_no_topic_arn_set():
@@ -102,7 +102,7 @@ def test_throws_keyerror_if_no_topic_arn_set():
     event = ingest_request(upload_url='s3://bukkit/example.zip')
 
     with pytest.raises(KeyError, match='TOPIC_ARN'):
-        start_ingest.main(event=event)
+        request_ingest.main(event=event)
 
 
 def ingest_request(upload_url, callback_url=None):
